Remove irc.js and require it instead
First off, don't poll this!
i just want you to test it out. it has some errors but it works and you don't have to change stuff in irc.js anymore (3 large files removed and it require irc.js instead) i know you said it would not work as you guys want but i can't see the problem and it works for me. this is part1. some little changes in client.js is made but i will check if i can make 1 pull with 2 files